# git-contextor/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Git Contextor API",
    "version": "1.1.6",
    "description": "A comprehensive API for Git Contextor - a code context tool with vector search, AI chat, and real-time monitoring capabilities. Git Contextor indexes your codebase and provides semantic search, AI-powered chat, and code sharing features.",
    "contact": {
      "name": "Git Contextor Support",
      "url": "https://github.com/stromdao/git-contextor"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3333",
      "description": "Local development server (default port)"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/api/search": {
      "post": {
        "summary": "Semantic Search",
        "description": "Performs a semantic search for contextually relevant code chunks using vector similarity.",
        "operationId": "semanticSearch",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["query"],
                "properties": {
                  "query": {
                    "type": "string",
                    "description": "Natural language query to search for",
                    "example": "how to implement user authentication"
                  },
                  "maxTokens": {
                    "type": "integer",
                    "description": "Maximum number of tokens to return in the response",
                    "example": 4096
                  },
                  "filter": {
                    "type": "object",
                    "description": "Optional filters to apply to the search"
                  },
                  "llmType": {
                    "type": "string",
                    "description": "Type of LLM to optimize context for"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "query": {
                      "type": "string",
                      "example": "how to implement user authentication"
                    },
                    "optimizedContext": {
                      "type": "string",
                      "description": "Optimized context string for LLM consumption",
                      "example": "--- File: src/auth/jwt.js (Score: 0.92) ---\\n..."
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SearchResult"
                      }
                    },
                    "tokenCount": {
                      "type": "integer",
                      "example": 1024
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable - configuration mismatch detected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat": {
      "post": {
        "summary": "AI Chat",
        "description": "Sends a query to the AI using repository context to generate an answer.",
        "operationId": "chatQuery",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["query"],
                "properties": {
                  "query": {
                    "type": "string",
                    "description": "Natural language query for the AI",
                    "example": "Explain the auth flow"
                  },
                  "context_type": {
                    "type": "string",
                    "description": "Type of context to use for the query",
                    "example": "general"
                  },
                  "include_summary": {
                    "type": "boolean",
                    "description": "Whether to include a collection summary in the response",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI response generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "query": {
                      "type": "string",
                      "example": "Explain the auth flow"
                    },
                    "response": {
                      "type": "string",
                      "description": "AI-generated response",
                      "example": "The authentication flow starts with a user providing credentials..."
                    },
                    "context": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SearchResult"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/reindex": {
      "post": {
        "summary": "Reindex Repository",
        "description": "Triggers a full re-indexing of the entire repository or a specific file. This is an asynchronous operation.",
        "operationId": "reindexRepository",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "description": "Optional specific file to reindex (relative path)",
                    "example": "src/auth/jwt.js"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File reindexed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Successfully reindexed file: src/auth/jwt.js"
                    }
                  }
                }
              }
            }
          },
          "202": {
            "description": "Full repository reindex started",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Full repository reindex started."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid file path",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Clear Index",
        "description": "Clears the entire vector index and collection data.",
        "operationId": "clearIndex",
        "responses": {
          "200": {
            "description": "Index cleared successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Collection and index data cleared successfully."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/share": {
      "post": {
        "summary": "Create Share",
        "description": "Creates a new share for public access to repository context.",
        "operationId": "createShare",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Share created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Active Shares",
        "description": "Lists all active shares.",
        "operationId": "listShares",
        "responses": {
          "200": {
            "description": "Shares retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "shares": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ShareResponse"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/share/tunnel": {
      "post": {
        "summary": "Start Tunnel",
        "description": "Starts a tunnel service to expose the Git Contextor server publicly.",
        "operationId": "startTunnel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["service"],
                "properties": {
                  "service": {
                    "type": "string",
                    "description": "Tunnel service to use",
                    "example": "localtunnel"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Tunnel service is starting",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Tunnel service 'localtunnel' is starting."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid service",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get Tunnel Status",
        "description": "Gets the current status of the tunnel service.",
        "operationId": "getTunnelStatus",
        "responses": {
          "200": {
            "description": "Tunnel status retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["stopped", "starting", "running", "error"],
                      "example": "running"
                    },
                    "url": {
                      "type": "string",
                      "nullable": true,
                      "example": "https://abc123.localtunnel.me"
                    },
                    "service": {
                      "type": "string",
                      "nullable": true,
                      "example": "localtunnel"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Stop Tunnel",
        "description": "Stops the active tunnel service.",
        "operationId": "stopTunnel",
        "responses": {
          "200": {
            "description": "Tunnel stopped successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Tunnel stopped."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/config": {
      "get": {
        "summary": "Get Configuration",
        "description": "Retrieves the current Git Contextor configuration.",
        "operationId": "getConfig",
        "responses": {
          "200": {
            "description": "Configuration retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Config"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Update Configuration",
        "description": "Updates the Git Contextor configuration. Changes require a service restart to take effect.",
        "operationId": "updateConfig",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Config"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Configuration updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Configuration saved. Please restart the service to apply changes."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/mcp/v1/spec": {
      "get": {
        "summary": "Get MCP Specification",
        "description": "Returns the Model Context Protocol (MCP) specification for this repository.",
        "operationId": "getMCPSpec",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "MCP specification retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "Git Contextor: my-repo"
                    },
                    "description": {
                      "type": "string",
                      "example": "Provides context-aware search for the my-repo repository."
                    },
                    "tools": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string",
                            "example": "code_search"
                          },
                          "description": {
                            "type": "string",
                            "example": "Searches the repository for code snippets, file contents, and documentation relevant to the user query."
                          },
                          "parameters": {
                            "type": "object",
                            "properties": {
                              "type": {
                                "type": "string",
                                "example": "object"
                              },
                              "properties": {
                                "type": "object"
                              },
                              "required": {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid Bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/mcp/v1/tools/code_search/invoke": {
      "post": {
        "summary": "Invoke MCP Code Search",
        "description": "Invokes the MCP code search tool to find relevant code snippets.",
        "operationId": "invokeMCPCodeSearch",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["query"],
                "properties": {
                  "query": {
                    "type": "string",
                    "description": "The natural language query to search for"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Code search results returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "content": {
                      "type": "string",
                      "description": "Formatted search results"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid Bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for authentication. You can get this from the UI config endpoint or from your Git Contextor configuration."
      },
      "ShareKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-share-key",
        "description": "Share key for accessing shared repositories."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token for MCP (Model Context Protocol) authentication. The token is the API key of the share."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
      "SearchResult": {
        "type": "object",
        "properties": {
          "filePath": {
            "type": "string",
            "description": "Path to the file containing the result",
            "example": "src/auth/jwt.js"
          },
          "score": {
            "type": "number",
            "description": "Relevance score (0-1)",
            "example": 0.92
          },
          "content": {
            "type": "string",
            "description": "Content of the code chunk"
          },
          "startLine": {
            "type": "integer",
            "description": "Starting line number",
            "example": 10
          },
          "endLine": {
            "type": "integer",
            "description": "Ending line number",
            "example": 25
          }
        }
      },
...
      "ShareRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "description": "Description of the share"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "Expiration date of the share"
          },
          "max_queries": {
            "type": "integer",
            "description": "Maximum number of queries allowed"
          },
          "scope": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Scope of the share (file patterns, etc.)"
          }
        }
      },
      "ShareResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the share"
          },
          "description": {
            "type": "string",
            "description": "Description of the share"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "Expiration date of the share"
          },
          "max_queries": {
            "type": "integer",
            "description": "Maximum number of queries allowed"
          },
          "access_count": {
            "type": "integer",
            "description": "Number of queries used"
          },
          "scope": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Scope of the share"
          },
          "api_key": {
            "type": "string",
            "description": "API key for accessing the share"
          },
          "url": {
            "type": "string",
            "description": "URL to access the share"
          }
        }
      },
...
    }
  },
...
}